requirements.txt
asgiref==3.5.2
certifi==2022.12.7
charset-normalizer==3.0.1
Django==4.1.4
django-ckeditor==6.5.1
django-js-asset==2.0.0
djangorestframework==3.14.0
idna==3.4
mysqlclient==2.1.1
Pillow==9.3.0
pytz==2022.7.1
requests==2.28.2
sqlparse==0.4.3
stripe==5.1.1
tzdata==2022.7
urllib3==1.26.14
